+++ CHANGELOG.md
@@ -7,6 +7,10 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## Version 0.2.4 - Unreleased
 
+### Added
+
+* Experimental `delayed_load` method for an image
+
 ### Fixed
 * Fixed `cats_per_img` in kwcoco stats.
 * Fixed `grab_camvid`.